fix review hints
