+++ builds/reference/src/index.ts
@@ -17,8 +17,8 @@ import { bindings as componentBindings } from '@tko/binding.component'
 import { filters } from '@tko/filter.punches'
 
 import components from '@tko/utils.component'
-import { createElement, Fragment } from '@tko/utils.jsx'
-import { JsxObserver } from '@tko/utils.jsx'
+import { createElement, Fragment, render } from '@tko/utils.jsx'
+import type { JsxRenderResult } from '@tko/utils.jsx'
 
 import { options } from '@tko/utils'
 
@@ -31,10 +31,7 @@ type ReferenceBuildExtensions = {
   jsx: {
     createElement: typeof createElement
     Fragment: typeof Fragment
-    render(jsx: any): {
-      node: ChildNode | DocumentFragment | null
-      dispose: () => void
-    }
+    render(jsx: any): JsxRenderResult
   }
   components: typeof components
   version: string
@@ -72,17 +69,7 @@ const referenceBuild: ReferenceBuildExtensions = {
   jsx: {
     createElement,
     Fragment,
-    /** Public render function that converts JSX to DOM nodes */
-    render(jsx: any) {
-      const fragment = document.createDocumentFragment()
-      const observer = new JsxObserver(jsx, fragment)
-      // Return the first child if single node, or the fragment if multiple
-      const node = fragment.childNodes.length === 1 ? fragment.firstChild : fragment
-      return {
-        node,
-        dispose: () => observer.dispose()
-      }
-    }
+    render
   },
   components,
   version,
