+++ packages/utils.jsx/src/index.ts
@@ -1,3 +1,4 @@
-export { getOriginalJsxForNode, maybeJsx, createElement, Fragment } from './jsx'
+export { getOriginalJsxForNode, maybeJsx, createElement, Fragment, render } from './jsx'
+export type { JsxRenderResult } from './jsx'
 
 export { default as JsxObserver } from './JsxObserver'
